+++ src/ModTemplate/_ModFrameworkDefinitions/Modules/ModFrameworkTypes.lua.d.lua
@@ -29,9 +29,9 @@ Types.ComponentTypes = ComponentTypes
 ---Component sizes
 ---@class ComponentSizes
 local ComponentSizes = {
-	--a single hanger slot
+	--a single hangar slot
 	Small = 1,
-	--a full hanger row, 5 slots
+	--a full hangar row, 5 slots
 	Large = 5
 }
 Types.ComponentSizes = ComponentSizes
@@ -300,14 +300,14 @@ local PilotActivities = {
 Types.PilotActivities = PilotActivities
 
 ---Hangar index identifiers
----@class HangerIndexes
-local HangerIndexes = {
+---@class HangarIndexes
+local HangarIndexes = {
 	component_type = 2,
 	item_index = 3,
 	logo = 5,
 	logo_index = 11
 }
-Types.HangerIndexes = HangerIndexes
+Types.HangarIndexes = HangarIndexes
 
 ---A non complete list of Virtual Keys
 ---@class VirtualKeys